Fix: 주식 리스트 레이아웃 수정
레이아웃이 주식 리스트에서만 보이도록 수정했습니다.
